requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
bleach[css]==6.1.0
    # via -r requirements.in
django==5.0.7
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
markdown==3.6
    # via -r requirements.in
packaging==24.1
    # via gunicorn
psycopg2-binary==2.9.9
    # via -r requirements.in
pygments==2.18.0
    # via -r requirements.in
python-frontmatter==1.1.0
    # via -r requirements.in
pyyaml==6.0.1
    # via python-frontmatter
six==1.16.0
    # via bleach
sqlparse==0.5.1
    # via django
tinycss2==1.2.1
    # via bleach
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
